+++ js/components/components.module.js
@@ -1,4 +1,5 @@
 angular.module('components', [
     'components.project',
+    'components.profile',
     'components.auth'
 ]);