+++ js/components/profile/profile-detail/profile-detail.component.js
@@ -1,0 +1,12 @@
+var profileDetail = {
+    bindings: {
+        feature: '<',
+        onUpdate: '&'
+    },
+    template: require('./profile-detail.tpl.html'),
+    controller: 'ProfileDetailController'
+};
+
+angular
+    .module('components.profile')
+    .component('profileDetail', profileDetail);