+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 * Made it possible to enter time entries in the past (if time is given)
 * Stats-bar for "types" are shown.
 * More component driven setup
-* Started with Project-component
+* Created project-component to save defaults
+* Created profile-component which allows to enable / disable features.
+
 
 ## 0.54 - 07.09.2017 - Refinements
 * Animation improvements